Finish change SelectedMeasurement menu (#35)
* Change to a version catalog

* Convert to using objects for navigation instead of strings for routes https://developer.android.com/develop/ui/compose/navigation

* Convert Big Decimal and Float to Double for Work

* Convert to using objects for navigation for AddWork and ExerciseList

* Add SelectMeasurement menu and implement logic on measurement change

+++ app/src/main/java/pro/selecto/slothos/data/entities/Work.kt
@@ -27,7 +27,7 @@ class Work(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "id") val id: Int = 0,
     @ColumnInfo(name = "order") val order: Int = 0,
-    @ColumnInfo(name = "value") val value: Float = 0F,
+    @ColumnInfo(name = "value") val value: Double = 0.0,
     @ColumnInfo(name = "name") val name: String = "",
     @ColumnInfo(name = "measurement_type") val measurementType: StandardMeasurementType = StandardMeasurementType.POUNDS,
     @ColumnInfo(name = "set_id") var setId: Int = 0,
